Reorganize `LlvmGenVisitor` implementation.
Implementation of this type is now split across multiple files
(corresponding to the AST types they generate code for) for
better organization.
